Do_Interaction_Backend
Also updated some glitches on the marketplace page. But now the interaction page will fill in, update timestamps and money when interactions are done. The interaction cooldown is 15 minutes. The selling form executes and changes with ease.  The user is redirected away from the page if the cat isn't theirs, the cat is invalid, or if the user isn't logged in.